rename: {name}page.ts -> types.ts로 파일명 변경 & path 적용
